JavaScript: Migrate the build from `rollup` to `rolldown`
