+++ javascript/packages/browser/package.json
@@ -18,8 +18,8 @@
     "build": "yarn clean && yarn build:wasm && yarn build:javascript",
     "build:wasm": "cd ../../../wasm && make ../javascript/packages/browser/build/libherb.js && cd -",
     "clean:wasm": "cd ../../../wasm && make clean_browser && cd -",
-    "build:javascript": "rollup -c",
-    "dev": "rollup -c -w",
+    "build:javascript": "tsc --emitDeclarationOnly && rolldown -c",
+    "dev": "rolldown -c -w",
     "clean": "rimraf dist && rimraf build",
     "test": "vitest run",
     "test:ui": "vitest --watch --ui",
