+++ javascript/packages/core/package.json
@@ -15,8 +15,8 @@
   "types": "./dist/types/index.d.ts",
   "scripts": {
     "download-html-entities": "node scripts/download-html-entities.mjs",
-    "build": "yarn download-html-entities && yarn clean && rollup -c",
-    "dev": "rollup -c -w",
+    "build": "yarn download-html-entities && yarn clean && tsc --emitDeclarationOnly && rolldown -c",
+    "dev": "rolldown -c -w",
     "clean": "rimraf dist",
     "test": "vitest run",
     "prepublishOnly": "yarn clean && yarn build && yarn test"
