+++ javascript/packages/formatter/package.json
@@ -18,8 +18,8 @@
     "herb-format": "bin/herb-format"
   },
   "scripts": {
-    "build": "yarn clean && rollup -c rollup.config.mjs",
-    "dev": "rollup -c rollup.config.mjs -w",
+    "build": "yarn clean && tsc --emitDeclarationOnly && rolldown -c",
+    "dev": "rolldown -c -w",
     "clean": "rimraf dist",
     "test": "vitest run",
     "test:watch": "vitest --watch",
