+++ javascript/packages/herb-language-server/package.json
@@ -18,8 +18,8 @@
   "module": "./dist/index.js",
   "types": "./dist/types/index.d.ts",
   "scripts": {
-    "build": "yarn clean && tsc -b && rollup -c",
-    "dev": "rollup -c -w",
+    "build": "yarn clean && tsc -b && rolldown -c",
+    "dev": "rolldown -c -w",
     "watch": "tsc -b -w",
     "clean": "rimraf dist",
     "test": "echo 'TODO: add tests'",
