+++ javascript/packages/language-service/package.json
@@ -14,8 +14,8 @@
   "module": "./dist/herb-language-service.esm.js",
   "types": "./dist/types/index.d.ts",
   "scripts": {
-    "build": "yarn clean && rollup -c",
-    "dev": "rollup -c -w",
+    "build": "yarn clean && tsc --emitDeclarationOnly && rolldown -c",
+    "dev": "rolldown -c -w",
     "clean": "rimraf dist",
     "test": "vitest run",
     "prepublishOnly": "yarn clean && yarn build && yarn test"
