+++ javascript/packages/node-wasm/package.json
@@ -19,8 +19,8 @@
     "build": "yarn clean && yarn build:wasm && yarn build:javascript",
     "build:wasm": "cd ../../../wasm && make ../javascript/packages/node-wasm/build/libherb.js && cd -",
     "clean:wasm": "cd ../../../wasm && make clean_node && cd -",
-    "build:javascript": "rollup -c",
-    "dev": "rollup -c -w",
+    "build:javascript": "tsc --emitDeclarationOnly && rolldown -c",
+    "dev": "rolldown -c -w",
     "clean": "rimraf dist && rimraf build",
     "test": "vitest run",
     "test:ui": "vitest --watch --ui",
