+++ javascript/packages/node/package.json
@@ -30,9 +30,9 @@
   },
   "scripts": {
     "install": "yarn vendor && node-pre-gyp install --fallback-to-build",
-    "dev": "rollup -c -w",
+    "dev": "rolldown -c -w",
     "build": "yarn clean && yarn templates && yarn vendor && yarn rebuild",
-    "build:javascript": "rollup -c",
+    "build:javascript": "tsc --emitDeclarationOnly && rolldown -c",
     "templates": "rake templates",
     "revendor": "yarn clean && yarn templates && yarn vendor",
     "vendor": "node ./bin/vendor.cjs",
