+++ javascript/packages/rewriter/rolldown.config.mjs
@@ -1,7 +1,3 @@
-import typescript from "@rollup/plugin-typescript"
-import { nodeResolve } from "@rollup/plugin-node-resolve"
-import commonjs from "@rollup/plugin-commonjs"
-import json from "@rollup/plugin-json"
 import { createRequire } from "module"
 
 const external = [
@@ -30,17 +26,7 @@ export default [
       sourcemap: true,
     },
     external: isExternal,
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        declaration: true,
-        declarationDir: "./dist/types",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
   {
     input: "src/index.ts",
@@ -50,15 +36,7 @@ export default [
       sourcemap: true,
     },
     external: isExternal,
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
 
   // Loader entry point (includes built-in rewriters and custom rewriter loader)
@@ -70,17 +48,7 @@ export default [
       sourcemap: true,
     },
     external: isExternal,
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        declaration: true,
-        declarationDir: "./dist/types",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
   {
     input: "src/loader.ts",
@@ -90,14 +58,6 @@ export default [
       sourcemap: true,
     },
     external: isExternal,
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
 ]
