+++ javascript/packages/stimulus-lint/rolldown.config.mjs
@@ -1,8 +1,3 @@
-import typescript from "@rollup/plugin-typescript"
-import { nodeResolve } from "@rollup/plugin-node-resolve"
-import json from "@rollup/plugin-json"
-import commonjs from "@rollup/plugin-commonjs"
-
 // Bundle the CLI entry point into a single CommonJS file.
 // Exclude Node built-in so they remain as externals.
 const external = [
@@ -35,16 +30,6 @@ export default [
       sourcemap: enableSourcemaps,
     },
     external: isExternal,
-    plugins: [
-      nodeResolve(),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        rootDir: "src/",
-        module: "esnext",
-      }),
-    ],
   },
 
   // Library exports (ESM)
@@ -56,16 +41,6 @@ export default [
       sourcemap: enableSourcemaps,
     },
     external: ["@herb-tools/core", "@herb-tools/highlighter", "@herb-tools/linter", "@herb-tools/node-wasm", "stimulus-parser"],
-    plugins: [
-      nodeResolve(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        declaration: true,
-        declarationDir: "./dist/types",
-        rootDir: "src/",
-      }),
-    ],
   },
 
   // Library exports (CommonJS)
@@ -77,15 +52,5 @@ export default [
       format: "cjs",
       sourcemap: enableSourcemaps,
     },
-    plugins: [
-      nodeResolve(),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        rootDir: "src/",
-        module: "esnext",
-      }),
-    ],
   },
 ]
