+++ javascript/packages/tailwind-class-sorter/rolldown.config.mjs
@@ -1,8 +1,3 @@
-import typescript from "@rollup/plugin-typescript"
-import { nodeResolve } from "@rollup/plugin-node-resolve"
-import commonjs from "@rollup/plugin-commonjs"
-import json from "@rollup/plugin-json"
-
 export default [
   {
     input: "src/index.ts",
@@ -21,17 +16,7 @@ export default [
       "path",
       "url"
     ],
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        declaration: true,
-        declarationDir: "./dist/types",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
 
   {
@@ -51,14 +36,6 @@ export default [
       "path",
       "url"
     ],
-    plugins: [
-      nodeResolve({ preferBuiltins: true }),
-      commonjs(),
-      json(),
-      typescript({
-        tsconfig: "./tsconfig.json",
-        rootDir: "src/",
-      }),
-    ],
+    platform: "node",
   },
 ]
