+++ javascript/packages/tailwind-class-sorter/tsconfig.json
@@ -13,6 +13,8 @@
     "isolatedModules": true,
     "preserveConstEnums": true,
     "outDir": "dist",
+    "declaration": true,
+    "declarationDir": "./dist/types",
 
     "skipLibCheck": true,
     "strict": true,
