+++ package.json
@@ -11,10 +11,6 @@
   },
   "devDependencies": {
     "@jothepro/doxygen-awesome-css": "https://github.com/jothepro/doxygen-awesome-css#v2.4.2",
-    "@rollup/plugin-commonjs": "^29.0.3",
-    "@rollup/plugin-json": "^6.1.0",
-    "@rollup/plugin-node-resolve": "^16.0.1",
-    "@rollup/plugin-typescript": "^12.1.4",
     "@typescript/native": "npm:typescript@^7.0.2",
     "@vitest/browser-playwright": "^4.1.8",
     "@vitest/ui": "^4.1.8",
@@ -24,9 +20,10 @@
     "nx": "22.7.7",
     "oxlint": "^1.77.0",
     "playwright": "^1.62.1",
+    "postcss": "^8.5.6",
     "prettier": "^3.9.6",
     "rimraf": "^6.1.3",
-    "rollup": "^4.62.3",
+    "rolldown": "^1.2.3",
     "rollup-plugin-postcss": "^4.0.2",
     "tsx": "^4.23.1",
     "typescript": "npm:@typescript/typescript6@^6.0.2",
